completely rewrote all doors code to condense them into two files - one with the nodebox models, and one with the actual node definitions.
In the process, cleaned up the code that handles node_ownership
and protection mods.

Also made the mod automatically enable either left-click or right-click to
open/close a door, depending on whether the game supports it (actually it
checks for the existance of an unrelated function that happened to be added at
the same time as the new right-click function).

Renamed some textures, duplicated all of the _lr.png textures to make it
easier to register the textures in a loop and to allow for more flexible
texturing.