Add ability to place a sign directly onto a brass/wrought iron fencepost
rather than crafting a sign-on-fencepost from one of each, how about just
placing a sign directly onto a fencepost instead?  This commit adds this.
Place a brass or wrought iron fencepost, then attempt to place a sign against
it.  The fencepost will be replaced with one bearing the sign - in the same
node.  Dig to get the two pieces back as separate items, right click the placed
sign to edit the text thereon.

Note that as of this commit, there is a bug in minetest/builtin/item.lua that
causes signs not to be deducted from your inventory on placement.