Added experimental alpha support to fancy door, disabled by default.
To enable, uncomment "use_texture_alpha" on lines 72 and 99 of
doors_and_gates.lua, and make sure that if you are using a texture pack,
that it has support for alpha on these doors. (HDX does, as do the
default textures).

Note that alpha support is glitchy - without shaders it will not work at all,
and the doors' nodeboxes will be screwed up.  With shaders, it mostly works,
but the whole door is slightly translucent where it should not be.