add blank canvas craft item and recipes for paintings
First, craft a blank canvas from four sticks
and some white wool:

- s -
s W s
- s -

Then, combine that canvas with some black and white dyes in the right
pattern to get the painting you want.

Paintings are numbred in-game, from 1 to 20.  Since there are so many
paintings, and so little crafting space, the only way I could think to
do it was to use binary for the recipe.  Convert the number of the
painting you want into binary, then use black dye for each 0 and white for
each 1, and lay them down into the crafting grid in this pattern:

bit2  bit1  bit0
  -   bit4  bit3
  -   Canv.   -

With the blank canvas in the bottom center slot.  For example, if you
want painting number 9, that would be 01001 in binary, so you'd arrange
the crafting grid like so:

Black Black White  --->  Painting #9
  -   Black White
  -   Canv.   -