add brass and steel chainlinks, recipe changes...
if glooptest is not installed, an alias is added to equate homedecor
steel chainlinks to glooptest's, for compatibility.

The six-steel-ingot wall chain recipe is no longer to make wall chains.
Instead, if glooptest is not installed, this recipe is used to make
chain *links*.

Whether glooptest is installed or not, the steel chainlinks are still
used along with steel ingots, to make wall chains as before.