Add a tree swing model and textures by jp
To place, point at the underside of a node, e.g. a ceiling or the bottom
side of tree leaves or something.  It will automatically stretch
downward trying to find the ground (well, anything that isn't
buildable_to anyway), down to a maximum of 4 meters' worth of rope.

To remove, dig the swing itself and it'll take the rope with it.