Make default and locked wooden signs old-style wallmounted instead of using 6d facedir
Caveat: this breaks worlds that used signs_lib, in the sense that many
(if not most) default and wooden locked wall signs in areas that were
affected by signs_lib will have to be fixed using the screwdriver.

The intent of this commit is to make signs_lib behave the same way
default signs always have in the past, instead of forcing 6dfacedir on
everyone.

This sucks, but it had to be done sooner or later.  when this mod was
first written, wallmounted nodeboxes weren't possible.

This change ONLY affects the default wooden wall signs and locked wooden
wall signs.  Metal signs, yard signs, signs on fenceposts, and hanging
signs are not affected.