convert lrfurn to use param2 colorization
To maximize the palette, this uses the colorwallmounted param2 style,
which necessitated modifying the models to make wallmounted look like
normal facedir.  This means couches and chairs can now only ever be
placed upright, facing one of N/S/E/W like a couch or chair normally
would, but since the screwdriver was already disabled, it was basically
impossible to rotate them to any other orientation anyway.

Since this is colorwallmounted, it uses the 32-color abridged Unified Dyes
palette.